+++ packages/api-client/package.json
@@ -2,7 +2,7 @@
   "name": "@tamanu/api-client",
   "version": "1.29.0",
   "private": true,
-  "description": "TODO",
+  "description": "API client for Tamanu Facility Server",
   "exports": {
     ".": "./dist/index.js"
   },
